+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+* Removed `black` from `style.lint` and `style.format`.
+
 
 ## [0.3.2] 2025-01-29